Rename game into Game
